Extract settings persistence/state helpers into taskpane-settings.js
Move LOCAL_SETTINGS_STORAGE_KEY, SETTINGS_COLLAPSED_KEY,
SETTINGS_CONTROL_CONFIG, DEFAULT_CALCULATION_SETTINGS, and helpers
readSettingsFromLocalStorage, saveSettingsToLocalStorage,
clearSavedLocalSettings, applySettingsToPanel, loadSavedSettingsIntoPanel,
applySettingsCollapseState, initializeSettingsCollapse out of taskpane.js
into a new dedicated module.

taskpane.js imports the seven public exports it still needs.
No behavior changes — architecture cleanup only.
